refactor: move dashboard components to dashboard sections
